fix(cli): Fix join+taint regression (#5846)
Join mode uses the output from Semgrep in a way that defeats the type
safety provided by ATD and mypy. When I split out the dataflow trace
types into a core version and a CLI version in #5765, I didn't realize
that join mode relies on the `extra` field of the CLI output being
compatible with the `extra` field of the core output, mypy was not able
to warn me, and there were apparently no tests exercising this behavior.

This changes the join mode code to explicitly convert the `extra` field
from the CLI match version to the core match version. Because this
conversion is now done explicitly, mypy will be able to flag
incompatibilities caused by future changes to the extra datatypes, but
there are still a lot of other places in this file where there is
minimal type safety.

At some point, it would be nice to:
* Add an automated test exercising join+taint mode
* Refactor so that join mode doesn't need to convert from the semgrep
output back to a datatype used earlier in the pipeline
* Refactor so that mypy can check the join mode code

Fixes #5839

Test plan: Automated tests, mypy, and manual testing of the issue as reported
in #5839