Fix lookup error for generated target files (#5848)
Currently the parsing statistics metrics creates a dictionary of targets
to help track the number of files which have errors. In the event that a
parse error is reported in a file not registered in the original plan,
however, this causes a KeyError due to a lookup failure in the
dictionary. This commit resolves this by ignoring parse errors reported
in generated files (which would originate either from a
metavariable-pattern or an extract mode rule) and simply discarding
them.

In the future we could restructure this to allow core to register
additional specific targets, or handle the parsing statistics directly,
but as it currently stands the cli doesn't have a way to determine what
these targets would be ahead of time, as this requires matching logic.