Prototype a utility for pattern parse stats
It can be frustrating to write rules because it's hard to get a pattern that
even works. Sometimes the pattern extension isn't supported (e.g. `$X. ...$Y`).
Sometimes you don't understand what is required for a pattern (and sometimes
we allow partial expressions only for specific languages). Sometimes there's
just a pattern parse error. This makes it hard for users to write rules and
especially for new users to learn the tool.

One idea I had was to at least enforce that the patterns that are supposed
to parse, do parse. The ultimate goal of this utility would be to generate
valid patterns from a file's AST, including ones that require substituting
ellipses or other semgrep extensions, and running them as rules to see how
many parse. This would be a lot of work though and would generate a LOT of
patterns unless we found a solution (one idea: only generate one for each
AST construct?).

In the meantime, this is a quick experiment that just extracts expressions
from a target AST and tries to use those as patterns. This already takes a
fair amount of time because currently a single bad pattern makes Semgrep stop
trying to parse a file, so I run each pattern as an individual rule. Also, a
significant proportion of expressions that don't parse are missing tokens; if
the expression were correctly dumped it would probably parse. That dmight also
be a useful thing to measure, though!

Anyway, wanted to put up the PR so people could know what I was talking about
and give ideas. It's entirely possible to me that this doesn't make sense as
an approach. I didn't even bother to put up a file-level comment yet.

To run:

```
semgrep-core -lang <LANG> -pattern_parsing_stats <file/dir>
```

e.g.

```
semgrep-core -lang python -pattern_parsing_stats cli/stubs/packaging/
```