AST_to_IL: Require fresh variables to have fake tokens
Previously we created fresh auxiliary variables with real tokens attached
to them. This makes it hard to distinguish the "real" code variables from
the "fake" auxiliary variables introduced during AST-to-IL translation.
Taint-tracking uses the token attached to a variable to decide whether to
include it in the taint trace or not, so here we force all auxiliary
variables to carry fake tokens.

Closes PA-1672

test plan:
Run this rule:

    rules:
      - id: test
        languages:
          - python
        message: bad
        mode: taint
        pattern-sinks:
          - pattern: '"bad" + ...'
        pattern-sources:
          - pattern: foo.bar.$ANYTHING
        severity: WARNING

on this .py file:

    def asdf():
        bad = "bad"
        bad + foo.bar.baz.f()

Check that there are no more weird intermediate variables being reported
in the taint trace.