Implement templated rules
What: Adds a new kind of file, a "template", which can be used to build a
"rule"

Why: rules often get repetitive. For instance, for a given framework, there
is usually a fixed set of sinks. These sinks get reused in every rule. When
the set is updated, each rule has to be updated. We want to be able to
factorize out those sinks.

How: We add a very simple syntax for templates and two extra fields, `uses`
and `ref`, to rules.

For templates: a template consists of a (declaration, contents) pair. The
declaration determines how the template may be invoked. When a template is
invoked, it is replaced with its contents, evaluating any variables. A
template file is a list of templates.

For rules: use `uses` to import a template. Use `ref` to call a template

Example:

Rule:
```
rules:
  - id: printing
    uses:
      - python_lib: python-lib.yaml
    patterns:
      - ref: python_lib.print()
    message: we printed!
    languages: [python]
    severity: ERROR
```

Template:
```
templates:
  - declaration: print()
    contents:
       pattern: print(...)
```

Under the hood, this works by converting the rules to jsonnet and then using
jsonnet to generate a json rule. In this case, the libsonnet template is:

```
{
   print() :: {"pattern":"print(...)"},
}
```

The jsonnet rule is:
```
local python_lib = import '/Users/emma/.semgrep/semgrep_templates/template_2_python-lib.libsonnet';
[{"id":"printing","languages":["python"],"message":"we printed!","patterns":[python_lib.print()],"severity":"ERROR"}]%
```

And the final json rule, compiled using jsonnet, is

```
[
   {
      "id": "printing",
      "languages": [
         "python"
      ],
      "message": "we printed!",
      "patterns": [
         {
            "pattern": "print(...)"
         }
      ],
      "severity": "ERROR"
   }
]
```