fix(python): Include start and end tokens of Python f strings
This ensures that the start and end tokens are included in the AST node
representing f strings. It's generally good practice to make sure that
all of the tokens that make up a specific construct are included, but in
particular this addresses an issue with autofix that led to the quotes
and `f` of an f string getting dropped.

Fixes #2995

Test plan: Automated tests