Fix Not_found exception if using extract rule
Due to a changes in 307f675cc1, the logic for making the rule table in
Run_semgrep changed. However, this relies on the assumption that
rule_pairs and list_of_rule_ids should contain the exact same
rules/rule_ids. When running with extract mode rules, this is
intentionally not the case.

This commit fixes this, by filtering out any unpaired rule_ids present
in these lists. These filtered rule_ids comprise the extract mode rules
which were used to target that file, but which should not be run against
it in search mode.