tainting: Restore ArgToSink findings after taint labels
With taint labels we stopped producing certain ArgToSink findings because
the `requires` label-formula of the sink always evaluates to false, if
the only taint we have is "polymorphic taint" coming from an argument.
Instead, the `requires` should always evaluate to true for arguments!

This is still not a proper implementation of taint labels for
DeepSemgrep, but at least it allows us to fix the regressions that
8f26fabe29a introduced.

Fixes: 8f26fabe29a ("tainting: Add experimental taint labels feature (#5725)")

Closes PA-1750

test plan:
DeepSemgrep tests pass