cli.py: cli: Adjust "type: ignore" for 'commands'.
At run time I can see that 'ctx.command.commands' is a dictionary.
However, that attribute is not present in the Click documentation, so
mypy is unhappy without this intermediate assignment and warning
suppression.

I do not know how to reconcile the conflict between this usage and the
documentation, nor why I see a mypy complaint that evidently the CI
build does not.