chore: reraise-timeout Semgrep rule
As suggested in #5980

This rule has 100 findings. Based on some spot checking, this seems to
be working as intended. However, I'm not sure whether we actually want
to enforce that for every `try`, we catch and reraise `Timeout`. Some of
the findings are on simple things like
`try Hashtbl.find ... with Not_found -> ...`. It would be annoying to
have to catch and reraise `Timeout` there, but maybe worth it to have
the guarantees of this rule? I'd love to hear what others think.

If people would like to move forward with this, I can go through and fix
the existing findings before committing this.