feat(js): Add switch case pattern parsing (#5980)
This also begins to fall back on the tree-sitter parser to parse
patterns, rather than extending the pfff parser to support this. For
now, the only semgrep constructs that the tree sitter parser can handle
are metavariables (because they are valid JS syntax already), ellipsis,
and switch cases. We will likely want to extend that, because in its
current state, using switch cases as top level patterns will restrict
users from using more advanced Semgrep constructs in the pattern.

We will also likely want to add support for `default:` as a pattern. The
current behavior is confusing: `default:` does parse as a top level
pattern, but it parses as a label. Because it's a valid program on its
own, we'll have to resort to some hack to interpret it as a default
switch case and match accordingly.

See related PRs for context:
- https://github.com/returntocorp/ocaml-tree-sitter-semgrep/pull/338
- https://github.com/returntocorp/pfff/pull/547

Test plan: Automated tests