refactor: Generic machinery to try multiple pattern parsers
In #5980, I introduced logic to first try the pfff parser when parsing
patterns for JS/TS, and then the tree sitter parser, rather than relying
solely on the pfff parser as we had been doing previously. Now I find
myself wanting to do this for Java as well, so it is time to make this
logic more generic.

Test plan: Automated tests