Simplify Dockerfile and fix Arch Linux failure
This closes #5715

The main issue was that WORKDIR does not honor the USER
directive on arch linux (it seems to do on macOS), so
the ocaml-tree-sitter-core directory itself (the '.') was
owned by root which prevented configure to create files
(e.g., tree-sitter-config.ml) in it.

By simplifying the Dockerfile and doing the COPY of
the parent dir, the ocaml-tree-sitter-core directory
is now owned by the user and so configure works.

I've also split some RUN command in 2 because they
were hard to read and I prefer to separate different
concerns in different RUN commands.

test plan:
docker build .