fix(java): `public Foo() {}` as a pattern (#6022)
The pfff parser rejects `public Foo() {}` as a pattern, even though the
grammar appears to allow it and even though `Foo {}` (without privacy
modifiers) is parsed correctly. There is already a hack in place to
address an LR(1) conflict for this rule, and I suspect that it has gone
wrong somehow. Instead of spending time refining the hack, I decided to
use tree-sitter as a fallback to parse Java patterns.

This updates semgrep-java to pull in
https://github.com/returntocorp/ocaml-tree-sitter-semgrep/pull/339,
which implements basic Semgrep pattern grammar extensions and top-level
constructors.

This also updates semgrep-rules to pull in
https://github.com/returntocorp/semgrep-rules/pull/2366.

Fixes #5558

Test plan: Automated tests