Write deep-semgrep is proprietary warning to stderr
We have a warning that deep-semgrep is proprietary. Since this warning
is on stdout, it interferes with JSON ingestion. Instead we will log it
to stderr.

Alternatives considered: only log the warning when `--json` is passed.
I can do this if someone feels strongly, it would just make the code
more complicated.

Test plan: `semgrep --config <anything> . --deep 2> deep_stderr.txt`