chore(java): Upgrade semgrep-java (#6087)
In particular, this pulls in https://github.com/returntocorp/ocaml-tree-sitter-semgrep/commit/278bae677938ccd03f33bcc0e09c98e3204d21a5

That commit allows `...` as a standalone statement. Previously, when
using the tree sitter parser for Java patterns, `...` in a statement
context would have required a semicolon. That commmit also adds tests
for various contexts in which someone might want to use `...`.

The tree sitter parser still is not typically used for pattern parsing.
It is only used as a fallback when the pfff parser fails.

Fixes #6051

Test plan: Automated tests