refactor: Add rendered_fix to semgrep-core output
This allows semgrep-core to compute autofixes and pass them to the CLI
for application, rather than the CLI computing the fix. This allows us
to use information exclusive to semgrep-core when rendering an autofix
to text. Specifically, we plan to manipulate the AST rather than
manipulating text directly.

Test plan:

* Existing automated tests
* Manual test: Replace the `None` with `Some` and an arbitrary string.
  Run Semgrep with an autofix rule and confirm that the written string
  appears in the target file instead of the real autofix.