chore: update to latest semgrep-rules and fix regressions
The code in Unit_engine.ml was not handling the new .fixed.xxx files
that has been added recently (by Pieter I think).

This is hopefully one of the last manual update of semgrep-rules.
I've started to make a new github action workflow to automate it
so we can catch problems ASAP.

test plan:
make test