fix: Use semgrep-core's rendered fix when constructing RuleMatch
Previously, the fix from semgrep-core would get applied, but the CLI
would display the autofix that it computed.

Doing this earlier in the pipeline means that the CLI displays the same
fix that it applies. It also just seems cleaner.

I don't think this needs a changelog entry because this code won't
actually be exercised until #6143 is merged, so it's fixing something
that no user could have experienced.

Test plan:

* Automated tests
* Manual test with test case given in #4964, along with the code in
  #6143. Verify that the CLI displays the correct autofix in the match
  result in addition to applying the correct autofix.