feat: add file renames to diffing logic (#6157)
The current diffing logic uses the list of files that have findings on the HEAD commit, and scans those files in the BASE commit. It then goes finding by finding and removes any findings that are found in the BASE commit from the HEAD commit, so that Semgrep only reports new findings. Semgrep uses a ci_unique_key (rule_id, file_path, syntactic_context, and index) to determine that two findings on BASE/HEAD are actually the same.

The problem with this, is that when users re-name files, the ci_unique_key changes because the file_path changes. To address this, in this PR semgrep starts to use git diff's ability to detect file renames, and substitutes the old file name in for the new file name when scanning the BASE commit.

This will prevent us from showing old findings as "new" in a diff scan when a developer is refactoring the code and renaming files.

Full scan behavior (without baseline_ref) is unchanged by this PR, so semgrep-app will still report these findings as "new" when a full scan is done and findings are uploaded to semgrep-app. This means the findings in renamed files will show up for security engineers, but will never block or spam developers. A follow-up PR in semgrep-app will address that, but it is out of scope for this PR.