test: Test autofixes in semgrep-core
This extends the existing semgrep-core test runner that we use to test
matches so that it can also test autofix. The format is pretty simple:
add a `.fix` file with the fix pattern and a `.fixed` file with the
expected contents of the target after fixes are applied.

There are existing end-to-end tests which test autofix on the CLI side.
Unfortunately, modifying them involves updating a large snapshot file,
and maintaining any large quantity of them would be onerous.
Additionally, it's not easy to adapt tests from our large existing test
suite in semgrep-core to also exercise autofix.

Semgrep's `--test` flag can also test autofix (#5190), but it has the
same problems as the existing autofix e2e tests for these purposes.