fix: Use better hash function in AST-based autofix
In some work-in-progress to support more kinds of nodes, I ran into an
issue where a difference in `e_range` led to a failure to look up a node
that should have been printed by lifting the original text. This fixed it.

Test plan: Automated tests, including those added in #6193