tainting: Track taint inside if's conditions and throw's (#6212)
We were only checking inside `NInstr` and `NReturn` nodes and we forgot
about `NCond` and `NThrow` ones! As a result, if there was a sink within
one of those two kinds of node, we missed it. We would also miss sources
that add taint to variables by side-effect.

This is the problem with using `| _ -> ...` !

Closes PA-1933

test plan:
make test # added one test