refactor(autofix): Use `result` for better debug logging
@aryx and @emjin both requested better debug logging in #6143, and I
also experienced the pain of trying to extend autofix to work with a
particular example without it. @emjin also suggested using `result` for
the printers. I initially pushed back on that, since manipulating
results does add some complexity over options. But, after some thought,
I ended up agreeing that this was the best way to add detailed debug
logging.

This is mostly a straightforward refactoring to plumb `result` through
to `Autofix.ml`. There, we convert back to `option` after logging the
error.

I arbitrarily chose `let/` to be the let binding operator for `result`.
I don't see any precedent for a binding operator for `Result.bind` in
the codebase. I'd be perfectly happy to choose a different one if anyone
has a preference. Perhaps if this becomes widely used we could move it
to `Common` like we recently did for `let*` as the binding operator for
`option`.

I decided to include the backtrace in the error message for printing.
It's hard to construct a good error message there. Obviously, we can't
print the text representation of the AST node in question -- if we could
we wouldn't need the error message. The autogenerated `show_*` functions
are great, but for large AST nodes they will lead to very verbose
output. Simply printing out the kind of node might not give enough
context for someone to figure out which node in a large tree is the one
with the printing issue. So, I decided to go with a backtrace. As noted
inline, we could miss entries due to tail call optimization. Hopefully
that's not a big deal.

Test plan: Automated tests plus manual testing of a few error cases.
Example output below:

```
[0.018  Info       Main.Autofix         ] Failed to render fix `bar * 2)`:
[0.018  Info       Main.Autofix         ] Failed to parse fix pattern:
[0.018  Info       Main.Autofix         ] Parsing.Parse_error
[0.018  Info       Main.Autofix         ] Raised at Parser_python._menhir_errorcase in file "src/pfff/lang_python/parsing/Parser_python.ml", line 15137, characters 8-18
[0.018  Info       Main.Autofix         ] Called from Stdlib__Fun.protect in file "fun.ml", line 33, characters 8-15
[0.018  Info       Main.Autofix         ] Re-raised at Stdlib__Fun.protect in file "fun.ml", line 38, characters 6-52
[0.018  Info       Main.Autofix         ] Called from Stdlib__Fun.protect in file "fun.ml", line 33, characters 8-15
[0.018  Info       Main.Autofix         ] Re-raised at Stdlib__Fun.protect in file "fun.ml", line 38, characters 6-52
[0.018  Info       Main.Autofix         ] Called from Parse_pattern.parse_pattern in file "src/parsing/Parse_pattern.ml", line 135, characters 18-62
[0.018  Info       Main.Autofix         ] Called from Autofix.parse_pattern in file "src/fixing/Autofix.ml", line 27, characters 9-51
```

```
[0.018  Info       Main.Autofix         ] Failed to render fix `bar($X * 2)`:
[0.018  Info       Main.Autofix         ] Did not successfully replace metavariable(s) in the fix pattern: $X
```

```
[0.018  Info       Main.Autofix         ] Failed to render fix `(1 + 3)`:
[0.018  Info       Main.Autofix         ] Could not print fixed AST:
[0.018  Info       Main.Autofix         ] Failed to print AST with hybrid printer. First error:
[0.018  Info       Main.Autofix         ] Node does not appear in the original target or fix pattern
[0.018  Info       Main.Autofix         ]
[0.018  Info       Main.Autofix         ] Second error:
[0.018  Info       Main.Autofix         ] Could not print AST:
[0.018  Info       Main.Autofix         ] Raised by primitive operation at Ugly_print_AST.print_fail in file "src/core/ast/Ugly_print_AST.ml", line 87, characters 28-44
[0.018  Info       Main.Autofix         ] Called from Ugly_print_AST.python_printer#print_expr_kind in file "src/core/ast/Ugly_print_AST.ml", line 125, characters 19-36
[0.018  Info       Main.Autofix         ] Called from CamlinternalLazy.force_lazy_block in file "camlinternalLazy.ml", line 31, characters 17-27
[0.018  Info       Main.Autofix         ] Called from Hybrid_print.with_fallback in file "src/fixing/Hybrid_print.ml", line 43, characters 6-17
[0.018  Info       Main.Autofix         ] Called from Autofix_printer.print_ast in file "src/fixing/Autofix_printer.ml", line 161, characters 8-35
[0.018  Info       Main.Autofix         ] Called from Autofix.render_fix in file "src/fixing/Autofix.ml", line 103, characters 6-120
[0.018  Info       Main.Autofix         ] Called from Stdlib__Result.bind in file "result.ml" (inlined), line 23, characters 36-39
[0.018  Info       Main.Autofix         ] Called from Autofix.render_fix in file "src/fixing/Autofix.ml", line 77, characters 4-1023
[0.018  Info       Main.Autofix         ] Called from JSON_report.unsafe_match_to_match in file "src/reporting/JSON_report.ml", line 167, characters 4-217
[0.018  Info       Main.Autofix         ] Called from JSON_report.match_to_match in file "src/reporting/JSON_report.ml", line 187, characters 9-34
[0.018  Info       Main.Autofix         ] Called from Common.partition_either.part_either in file "src/pfff/commons/Common.ml", line 978, characters 15-18
```