refactor(test): Move autofix application code to Autofix.ml (#6223)
As requested in review of #6193. Currently this isn't suitable for use
in production. I've added todos in the places that I think may need to
be changed if we start using this in production, instead of just as part
of a test engine.

I also refactored a little bit so that we don't have to explicitly pass
the fix around separately. It's available in each pattern match, so now
we pull it from there as we are applying fixes.