AST_to_IL: Fix translation of lambdas (#6240)
Translating lambdas requires a recursive call to the
`function_definition` translator, and that requires a fresh `stmts`
stack, otherwise weird things can happen.

This was discovered by accident, and we are not aware of this causing
any bad result. As the analyses are implemented now, the incorrect
translation still "works" because the statements are still visited in
the same order. Now, if we change the translation (which I was trying
to do when I discovered this), then it may cause bad results.

test plan:
% cat bug.js
function test() {
  foo(bar(), (x) => { return x });
}
% bin/semgrep-core -lang js -dump_il bug.js
  # or
% bin/semgrep-core -lang js -cfg_il bug.js
  # now the IL representation is correct, `bar()` is followed by
  # `(x) => { return x }` in the IL, rather than `bar()` being
  # incorrectly nested inside `(x) => { ... }` as if the code were
  # `(x) => { bar(); return x }`.