refactor(autofix): Split expr printing into different methods
Instead of having the pattern match on expr_kind in each language
printer, we can have it in the base printer and then dispatch to the
methods for each variant. This is what the autogenerated visitors that
ppx deriving visitors does, and I think it makes the code more concise
and readable overall.

Test plan: Automated tests