jsonnet: support nested import
Before this PR, if you had:
a.jsonnet importing x/x.jsonnet which was importing from there
y/y.jsonnet (which final path is then x/y/y.jsonnet), then
semgrep was failing at the second import. This is because
we need to handle the base argument in import_callback.

test plan:
in deep-semgrep, the use of a semgrep.jsonnet including
another semgrep.jsonnet which includes other files now work.