test(autofix): Unit tests for autofix printer (#6263)
This tests the autofix printer to make sure that it is lifting the
original text from the target and fix_pattern where appropriate. Details
are in the main comment in `Unit_autofix_printer.ml`.

This also refactors Unit_engine to pull out the code that runs Semgrep
matching over some code and rule, so that it can be used here.

Test plan: Picked an arbitrary semgrep-core test and modify it to change
the matching behavior. Ensure that `make test` fails. Change one of the
new unit tests so that the output doesn't match. Ensure that `make test`
fails.