[RFC] Beginning of autofix printing statistics
The goal here is to record autofix printing statistics, like we do
parsing statistics. We want to look over rules repositories (starting
with semgrep-rules), run all the rules that have both a `fix` field and
a test case, and measure what fraction of them semgrep-core is able to
render a fix for.

I'm requesting feedback on this early because I'm uncertain of my
decision to implement this in Bash. I thought it would be a quick and
dirty script but it quickly became challenging to write and read. Maybe
that's okay since it's not likely to change very often, and isn't part
of production, but I'm not sure.

Should I:

1. Finish this Bash script, add comments, and improve names, then add a
   Python script like `parsing-stats/upload-parsing-rates` to upload the
   statistics
2. Rewrite this in Python
3. Put as much of this in OCaml as I can, like is done for parsing stats

I am leaning towards (1) or (2) but I would appreciate feedback