test: Autofix printing stats
This sets up a tool that clones rules repositories, finds the rules that
have a `fix:` field, runs them against their test cases, and counts how
many semgrep-core can succesfully render using AST-based autofix.

Optionally, it can also upload this data to our metrics dashboard.

This will allow us to track the performance of AST-based autofix on
real-world rules.

I leaned heavily on the parsing-stats code when creating this. The part
that clones the repositories is pretty much a direct translation from
Bash into OCaml.

I wrote this in the style of a shell script, making use of
commonly-available tools where appropriate. Down the line, we may want
to replace some of these with various OCaml utilities, but for now I
think this is reasonable.

I wrote this in OCaml after the discussion on
https://github.com/returntocorp/semgrep/pull/6276. My initial attempt to
write it in Bash did not turn out especially well.

I will follow up to create a CI job that runs this daily.

Test plan:

Manually run, both with and without the `--upload` flag. Without, see
that the stats for Python are printed. With, see that the stats for
Python are printed and
https://dashboard.semgrep.dev/metric/semgrep.core.python.autofix.pct is
updated.

Verify that subsequent runs without removing the `tmp` directory are
successful as well.