fix(autofix): Abort on failure to replace metavar anywhere in string literal
This is an extension of #6242. What I didn't realize when I wrote that
was that textual autofix allows a metavariable to appear anywhere within
a string literal. This makes sense, since the implementation can't
distinguish between the inside of a string literal and anything else.

Now, after performing metavariable replacement, we search string
literals for any instance of metavariables, and we abort if we find one.
Previously, we only aborted if the string literal's contents was,
exactly, the string literal in question.

Test plan: The test plan of #6242, then repeated with the fix changed to
`bar("x$Xy")`

Also repeat the test with the pattern `foo("$X", "$Y")`, the target
`foo("asdf", "bar")`, and the fix `bar("$Y", "$X")` to test that the
regular expression is joined properly.