feat(autofix): AST-based autofix for JS
All of the printing I have implemented so far for Python is common to
many languages, including JS. So, I introduced a `common_printer` class.
This is separate from `base_printer` as explained in the comment.

Then, I used that in both the Python printer and JS printer. The rest is
just updating a bit of boilerplate to include JS and adding tests.

Test plan: Automated tests. Also, run `autofix-printing-stats/run` and
verify that the Python output is unchanged, and JS reports 0.0% (0/1)
for the printing stats (there is only one JS fix in semgrep-rules, and
it is a statement which is currently out of scope for this project).