chore(autofix): Debug logging in CLI
Now, with the `--debug` flag, the Semgrep CLI will report which autofix
rendering strategy was used: AST-based autofix from semgrep-core, or
text-based autofix rendered in the CLI. This should help advanced users
experimenting with AST-based autofix, and should also help us debug
autofix issues when reported by users.

Test plan: Manual test with the `--debug` flag against the example given
in #4964. Verify that it is rendered correctly with AST-based autofix
and that the debug log reports that fact. Modify `Autofix.ml` to always
return `None`, and try again. Observe that the CLI reports that
text-based autofix was used, and that the extraneous commas reported in
that issue appear in the autofix.