chore: Add --verbose flag to autofix printing stats script
This will print a list of rule-target pairs that had at least one match
where an autofix failed to render.

I also refactored how flags are handled as part of this.

Test plan:

Ran without any flags and it behaved as before. Ran with `--verbose` and
got a list of failing test cases. Ran with `--verbose --upload` and got
"Uploading statistics" at the beginning (commented out the actual upload
so as not to pollute the metrics data).