fix(autofix): Handle focus-metavariable matches in AST-based autofix (#6381)
At the point where we try to render an autofix, the rule that we have
when there is a focus-metavariable is the one that has been created by
`Match_env.fake_rule_id` called from
`Match_search_mode.apply_focus_on_ranges`. This fake rule ID is later
adjusted in `Range_with_metavars.range_to_pattern_match_adjusted`.

I've updated that later adjustment to copy over the fix and the
languages, which are both needed for autofix.

Previously, the fix and languages weren't copied over either in the
place where the rule ID was created, or when adjusted later. I'm not
sure of the history or the broader architecture here, so I'm not sure
why we create a rule ID with dummy data and then update it later, rather
than just using the correct info to begin with. But, I think that my
changes fit in with the existing design. It might be time to update some
names, but I don't think I have the context to suggest improvements.

Test plan:

`./autofix-printing-stats/run` --> 96.4% (27/28) for Python, up from
67.9% (19/28)

Manually run the CLI against the python/distributed/security.yaml rule
in semgrep-rules, and verify that it performs the fix correctly.