refactor(autofix): Refactor fix application logic
This updates the fix application code, which is currently used only in
tests, to have pieces which are usable in production. This will be used
in the OCaml port of the CLI code in `autofix.py`.

In particular:
- I introduced the `textedit` type, which describes a single change to a
  file.
- Using `textedit` as an intermediate type, I separated the code that
  determines what changes to make based on a set of pattern matches from
  the code that actually applies the changes.
- I updated the code to abort on overlapping fixes to instead discard
  overlapping fixes, and report to callers which ones were discarded.

Note that instead of tracking changes to files like we do in
`autofix.py`, we simply apply fixes starting from the end of the file.

Test plan: Automated tests